requirements.txt
asttokens==2.4.1
black==24.4.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
flake8==7.0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-pptx==0.6.23
pytz==2024.1
pywin32==306
pyzmq==26.0.3
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
XlsxWriter==3.2.0